Additional note for release build
If org.gradle.configureondemand=true is set for gradle, gradle will skip bundleJsAndAsset gradle task, which will result into the app crashing on startup with this in the logs:

java.lang.RuntimeException: Unable to load script from assets 'index.android.bundle'. Make sure your bundle is packaged correctly or you're running a packager server

Since it is not clear from the error what is causing this, I would recommend adding this additional note to potentially save a lot of time for people figuring this out (it was not trivial to find the answer by googling).