Small clarification in Native Modules (iOS)
I'm an iOS beginner and missed this import. I later figured it out, but hopefully this PR will save others some time.

The import is correctly added in the next code sample below but I believe we should add it right in the first example - this is the first bit of code people copy&paste and it doesn't work currently.