Docusaurus provides us with an easy way to keep snapshots of the docs for each version of React Native. As a result, people can go back and see what a View component looked like in any past React Native release. (#466)
This is less relevant for "evergreen" content such as the Getting Started
guide, the Tutorial, and basically the rest of the Guides. These docs should
be getting updated across all versions at the same time. This script makes
sure we are not tracking separate versions of a unversioned doc.

We achieve this by ensuring any changes made to these files in `docs/` are
then persisted to the `versioned_docs/version-0.5` folder, as Version 0.5
is the oldest document version. This makes it so that the same text is
displayed to the reader, regardless of what version they are looking at.

This script is not yet invoked in CI or as a pre-commit hook, but the plan is
to enable it soon.