Create Updated blog post on how to use TypeScript
TypeScript is getting ever more popular. Which is why I took the time to updated and enhance the existing blog post about using React Native with TypeScript. This update is more comprehensive that the few lines that have been added to the old post a couple of days ago.

In addition to getting TypeScript and Jest to work, it also covers TSLint, Enzyme and React Native Testing Library.