Add warning about multiline centering the text on Android
React Native has this issue where when using a `TextInput` with `multiline` set as `true`, the text is aligned to the top on iOS and to the center on Android by default, which causes confusion, evidenced by the amount of likes in [this comment](https://github.com/facebook/react-native/issues/13897#issuecomment-300735822). I believe the `textAlignVertical` default should be changed when `multiline` is `true`, but since I don't know if there are any reasons the issue I linked was closed, I suggest this is at least documented.