Update fast-refresh.md (#1500)
* Update fast-refresh.md

Hi there!

As a library author I found it puzzling how fast-refresh affects the hooks. It is nice do document how **useState** and **useRef** differs from the others. Also giving an example of how **useEffect** runs related to a fast-refresh. Because it runs as if the component was removed and added again, which is just nice to know 😄 

Awesome work on this!

* Update fast-refresh.md

Grammar and stylistic changes

* Rewrite

Co-authored-by: Rachel Nabors <rachelnabors@users.noreply.github.com>
Co-authored-by: Dan Abramov <dan.abramov@gmail.com>