Modify duplicate code block in intro-react.md
There are two examples placed one on top of the other where the code is exactly the same. 

This makes the docs harder to understand, as referencing the same exact code for pointing out two different things in consecutive paragraphs can be confusing to a newcomer. It feels like a mistake, even though it can be backed that it is not.

I suggest removing the `export default Cat;` from the first example, where introducing the concept of default exports is not necessary, and use it only on the second code block, where the complete function component is displayed and once the concept has been properly introduced in the previous paragraph. 

I think this makes more sense and avoids the duplicate.