Minor wording changes in PlatformColor.md
I found this section specially hard to read, so I'm proposing some minor tweaks that I think will make it easier to understand. 

I also added a mention to the fallback values which the function can receive, taking from the original PR which introduced the changes (https://github.com/facebook/react-native/pull/27908)

I don't know what the exact policy is on this, but since this was introduced in RN 0.63.0, perhaps this could also be clarified on the docs, as using it from an older version will throw a `_reactNative.PlatformColor is not a function` error.