Update layout-props.md
It sounds awkward to say that `flex` in React Native works according to "the Yoga". According to the [Yoga homepage](https://yogalayout.com/), Yoga is "a highly optimized open source layout engine". Accordingly, it would be clearer to the reader if we change the language to say that `flex` works according to "the Yoga layout engine".