KeyboardAvoidingView: Describe `keyboardVerticalOffset` more precisely.
A common use case is to pass the height of a header, if the
`KeyboardAvoidingView` is just below the header without including
it. So, help make it easier to decide whether or not to use this
prop by clarifying what it does. When I read "the react native view"
and "some use cases", I had a vague idea that it probably wouldn't
help in my case, and that it was something I'd reach for if I was
using React Native for parts of my app and not others.

But it turned out to be quite useful for my use case!