Update intro-react-native-components.md
While I was learning this it looked to me like Native components and Core components are "siblings" from the way sub headings were put. The idea of "Native Components" being parent of "Core Components" and "Community Components" was not coming out (IMHO). I've just tried to re-arrange text to convert this idea with as minimum changes as possible.