Potentially confusing extraData flatlist docs
The current example for using extraData in the flatList does not actually require the extraData prop at all. This commit is to contrive an example where extraData in the flatList makes a difference.

The existing example runs fine without the extraData prop at all because the functional component re-creates the renderItem function reference on every call to the App function when the selectId state is set. Since the renderItem reference changes to the flatList, it re-renders. In the functional paradigm data needs to be passed to the renderItem method by reference in order to create the need for extraData.