Update handling-text-input.md to emphasize that setText(t) does not take component state for input
Very minor: Since `PizzaTranslator` has `text` as a variable in its state, I propose that the `onChangeText` callback use a different variable name to emphasize that this value is *not* derived from component state but instead is used to set it. I use `t` to avoid verbosity, but `newText` might also be appropriate.

Note: I am a React Native neophyte; if the above goes against stylistic conventions with this framework, then perhaps an explanatory note below this page's code example would suffice.