Adding a command
While starting the server , I found 'docusaurus' is not recognized as an internal or external command, operable program or batch file .  It took me some time to find the solution and found this command on stackoverflow. If included here, the command might help the user in dealing with such problems and it will save the user's time.