Update navigation alternative in navigation.md
The navigation router is an up and coming navigation library from @grahammendick which brings scene based navigation from native apps to React Native with 100% native navigation for both iOS and android. 

The benefits include: 

1) No more hardcoded urls in your code
2) No more having to manually keep a link config in sync with all your navigation screens.
3) No more performance problems with nesting stacks.
4) Works just as well on React as on React Native
4) Everything is 100% native

React-native-navigation from wix is a great library in its own right.

When new alternatives pop up it is our duty as open source developers to bring exposure to them. Let's keep the React Native ecosystem as dynamic and vibrant as we can and improve everybody's experience!

Any feedback is welcome and greatly apppreciated.