rename view#nativeid to view.md#nativeid
Small change, can we reference the Markdown file with extension? This helps a little with maintainability as these references can easily be searched for and Docusaurus should find and substitute the correct URL if it changes in future :).

(We use a mixed pattern currently but this approach is better than the "from root" /docs/view#nativeid approach I suggested originally.)
https://github.com/facebook/react-native-website/pull/3249#discussion_r946598670