Lint/Format all the examples
This adds a script which can run ESLint on each Snack example in Markdown. This is wired to the default RN ESLint configuration, and a Prettier configuration to limit to 70 lines to prevent wrapping. Apart from snforcing consistency, it means copy/pasting examples into a new RN project will not create formatting errors.

Not all examples are clean of lint errors/warnings so this is not enforced yet. The script is also currently very slow. But it's a good start, and we can later apply something similar to `tsc`.