Add optional "ext" parameter to SnackPlayer examples
This makes a couple changes in preparation for TypeScript examples. Going through 20 pages for core components, most examples can be shifted to be both JavaScript and strict TypeScript. Some are not compatible as both, and need to be separated as JS examples vs TS examples.

**Adds a parameter to SnackPlayer blocks for extension**
Snack defaults a codeblock to represent a file named [`App.js`](https://github.com/expo/snack/blob/main/docs/url-query-parameters.md#parameters). This extenstion effects how the file is parsed during bundling, and the syntax highlighting it is given in the editor. We allow passing the extension, so that specific examples can specifically target JS vs TS. This directive will also be used to determine how the file is used during linting.

**Change the default extension of SnackPlayer examples to `.tsx`:**
This is needed for Snack to parse an input as TypeScript during bundling and syntax highlighting. Vanilla JS is still allowed, and it seems to create an identical result to before in the snack player.