Update TypeScript Page and References to TypeScript
This change updates the main TypeScript page with new instructions for 0.71, where TypeScript is the default language, and RN types are included by default.

This does not yet update samples for TS, which is still a work in progress. I also chose to omit information about the 0.70 to 0.71 migration, since it seems like it might make more sense in either the release announcement or a separate blog post instead of long-term in the documentation, but I would be open to feedback on that.