Fix rendering of snacks on mobile
Switching pages on mobile currently causes a blank snack example to show.

This is root caused to a workaround on the `onRouteUpdate` to read the DOM after mutation, which seems to work on desktop, but not on the mobile version of the site. This change moves to a newer lifecycle method called after the document is rendered to the DOM, which fixes the issue.