Update Inline JSX and Type Definitions (1/2)
This is a best effort attempt at updating all of the non-snack code examples to TypeScript, adding type definitions to signatures, updating existing type signature documentation to align to TS instead of Flow, and applying a couple formatting changes for consistency with RN's prettier config.

A lot of instances of `removeListener` have been removed in code, and are now removed on the website. I also removed some APIs which should not be called by user code (e.g. functions called by native from the framework).

I also marked a couple of the removed APIs as removed instead of deprecated, and didn't bother to update them.

Finally, we now consistenyl end function declarations with a semicolon where before some docs did, and others didn't.