Update optimizing-flatlist-configuration.md
Based on my understanding, to gain the benefit of not having the `renderItem` function not recreate/rerun, it should be wrapped in a `useCallback` in functional components. The code snippet already demonstrates this. However, the docs do not explain this nor explain why this should be done.

Minor clarity adjustments as well. Let me know if my understanding is incorrect or if I should make further changes.