Specifcy Explicit React Native Version For Init (#3574)
`npx react-native init` [is not guaranteed](https://github.com/npm/cli/issues/4108) to pull in `react-native@latest`. When the bootstrapping CLI is mismatched from the react-native version it tries to install, it may not work correctly.

Currently, this means if you previously installed RN 0.70, you may end up using `react-native init` from RN 0.70, using CLI `9.2.1`, to bootstrap your React Native 0.71 app. This seems to lead to the CLI not correctly generating the template app. Most visibly, files like `_rbenv` are not converted to `.rbenv`.

This seems like a CLI bug. `--version` exists to allow CLI to bootstrap different RN version, so it should support this sceanrio. But we can change the example to one that prevents users from hitting it.