Android Java build fail
file  CenteredText.java in class name is CenteredText. However, in the CenteredTextManager.java file, we used the name RTNCenteredText instead of CenteredText. This is causing issues when trying to create a fabric component, as the build fails when updating the class name to CenteredText. Everything works fine when using the correct class name.