Naming inconsistency causing build failure in Android Java file. (#3678)
* Android Java build fail

file  CenteredText.java in class name is CenteredText. However, in the CenteredTextManager.java file, we used the name RTNCenteredText instead of CenteredText. This is causing issues when trying to create a fabric component, as the build fails when updating the class name to CenteredText. Everything works fine when using the correct class name.

* Update pillars-fabric-components.md

Naming inconsistency issue fixed in 0.70.0 version

* Update pillars-fabric-components.md

Naming inconsistency updated version 0.70.0

* Update pillars-fabric-components.md

Naming inconsistency updated name

* Update pillars-fabric-components.md

VERSION 0.70.0 there is classname itself RTNCenteredText so We don't need to update