Open twitter url in new tab (#4055)
The twitter url in hero section of home page is an external url, so clicking on it spoils the SPA router and opens the page in the same tab. This is going to give better user experience if this url opens in a new tab, and that's what I have done in this commit.